Move compilers to libocto/compilers subdir + clean up unnecessary alias
