+++ source/libocto/compilers/compiler-interface.ts
@@ -1,9 +1,9 @@
-import type { Result } from "../result.ts";
-import type { Agent, AssistantMessage, LoweredIR } from "./llm-ir.ts";
-import type { LoadedTools } from "./tool-def.ts";
-import { QuotaData } from "../utils/quota.ts";
-import { JsonFixResponse } from "../prompts/autofix-prompts.ts";
-import { Transport } from "../transports/transport-common.ts";
+import type { Result } from "../../result.ts";
+import type { Agent, AssistantMessage, LoweredIR } from "../llm-ir.ts";
+import type { LoadedTools } from "../tool-def.ts";
+import { QuotaData } from "../../utils/quota.ts";
+import { JsonFixResponse } from "../../prompts/autofix-prompts.ts";
+import { Transport } from "../../transports/transport-common.ts";
 
 export type CompilerIR<A extends Agent<any, any, any>> = LoweredIR<A["tools"]>;
